+++ works/views.py
@@ -9,34 +9,36 @@ from users.models import MyUser
 
 class ToDoListView(APIView):
     def post(self, request):
-        user = get_object_or_404(MyUser, pk=request.user.id)
-        print(user)
         serializer = ToDoSerializer(data=request.data)
         if serializer.is_valid():
-            #serializer.save(commit=False)
-            #serializer.author = user
-            serializer.save()
+            serializer.save(user=request.user)
             return Response({"message":"할일 저장"}, status=status.HTTP_201_CREATED)
         else:
             return Response({"message":serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
         
     def get(self, request):
-        works = ToDoList.objects.all()
+        works = ToDoList.objects.filter(user=request.user.id)
         serializer = ToDoSerializer(works, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
 
 class ToDoListDetailView(APIView):
-    def put(self, request, work_id):
+    def put(self, request, work_id): 
         work = get_object_or_404(ToDoList, pk=work_id)
-        serializer = ToDoSerializer(work, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+        if work.user == request.user:
+            serializer = ToDoSerializer(work, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"message":"유저가 달라요."}, status=status.HTTP_400_BAD_REQUEST)
         
     def delete(self, request, work_id):
         work = get_object_or_404(ToDoList, pk=work_id)
-        work.delete()
-        return Response({"message":"삭제완료"}, status=status.HTTP_200_OK)
+        if work.user == request.user:
+            work.delete()
+            return Response({"message":"삭제완료"}, status=status.HTTP_200_OK)
+        else:
+            return Response({"message":"유저가 달라요."}, status=status.HTTP_400_BAD_REQUEST)
